add: Routing for all pages
Route Group named (content-pages) added to use layout for required pages
app/page.jsx redirets users to /get-started
Fixes #23
