+++ app/(content-pages)/courses/[courseId]/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function page({ params: { courseId } }) {
+    return (
+        <div>{courseId}</div>
+    )
+}
