+++ app/(content-pages)/home/page.jsx
@@ -1,0 +1,11 @@
+"use client"
+
+import React, { useEffect } from 'react'
+
+export default function Home() {
+    return (
+        <div>
+            <h1>Home</h1>
+        </div>
+    )
+}
