+++ app/(content-pages)/layout.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function Layout({ children }) {
+    return (
+        <>
+            <navbar>
+                <a href="/">Home</a>
+            </navbar>
+            {children}
+        </>
+    )
+}
