+++ app/(content-pages)/my-learning/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function MyLearning() {
+    return (
+        <div>MyLearning</div>
+    )
+}
