+++ app/(content-pages)/profile/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Profile() {
+    return (
+        <div>Profile</div>
+    )
+}
