+++ app/(content-pages)/saved/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Saved() {
+    return (
+        <div>Saved</div>
+    )
+}
