+++ app/(content-pages)/search/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Search() {
+    return (
+        <div>Search</div>
+    )
+}
