+++ app/(content-pages)/statistics/page.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Statistics() {
+    return (
+        <div>Statistics</div>
+    )
+}
