+++ app/get-started/page.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import BrandIcon from '../components/BrandIcon'
+import TitleAndParagraph from '../components/TitleAndParagraph'
+
+export default function Welcome() {
+    return (
+        <main>
+            <BrandIcon />
+            <TitleAndParagraph />
+        </main>
+    )
+}
