--- app/page.js
@@ -1,11 +1,0 @@
-import BrandIcon from './BrandIcon'
-import TitleAndParagraph from "./components/TitleAndParagraph";
-
-export default function Home() {
-  return (
-    <main>
-      <BrandIcon />
-      <TitleAndParagraph />
-    </main>
-  );
-}
