+++ app/page.jsx
@@ -1,0 +1,4 @@
+import { redirect } from 'next/navigation'
+export default function GetStarted() {
+  redirect("/get-started")
+}
